Reverted fix for bug #4689 (closes #4987): Fix seems to cause problems, reinvestiation is needed
git-svn-id: https://svn.typo3.org/TYPO3v4/Core/branches/TYPO3_4-0@2033 709f56b5-9817-0410-a4d7-c38de5d9e867

+++ t3lib/class.t3lib_tcemain.php
@@ -1776,7 +1776,7 @@ class t3lib_TCEmain	{
 						}
 					}
 					$theDec = ereg_replace('[^0-9]','',$theDec).'00';
-					$value = (float)(intval(str_replace(' ','',$value)).'.'.substr($theDec,0,2));
+					$value = intval(str_replace(' ','',$value)).'.'.substr($theDec,0,2);
 				break;
 				case 'md5':
 					if (strlen($value)!=32){$set=false;}
